feat(discoveryRepositorySelection): styled buttons
